backend/training/requirements.txt
torch
torchvision
transformers
tqdm
Pillow
roboflow
PyYAML
